requirements.txt
helium
pandas
faker
bs4
cpca
selenium
transformers
moviepy
pillow